+++ examples/extension-projects/knowledge-distillation/src/stochaflow_knowledge_distillation/__init__.py
@@ -1,1 +1,0 @@
-"""Knowledge-distillation reference extension package."""